🚚 Rename: product -> products로 파일명 변경
